requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
billiard==4.1.0
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.1
cryptography==41.0.3
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.25.1
executing==1.2.0
Faker==19.6.0
flower==2.0.1
fonttools==4.42.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.8.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydyf==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
pyphen==0.14.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
ruff==0.0.289
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
stripe==6.6.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
weasyprint==60.0
webencodings==0.5.1
zopfli==0.2.3
